Fixed comments about mvn clean test error and moved data from dataProvider class
