+++ HomeWork1/pom.xml
@@ -21,7 +21,7 @@
             <artifactId>HomeWork1</artifactId>
             <version>1.0</version>
             <scope>system</scope>
-            <systemPath>${project.basedir}/src/main/resources/calculator-1.0.jar</systemPath>
+            <systemPath>${project.basedir}/src/test/resources/hw1/calculator-1.0.jar</systemPath>
         </dependency>
     </dependencies>
